% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpdir.R
\name{setTmpDir}
\alias{setTmpDir}
\title{Set the location of the temporary R session directory}
\usage{
setTmpDir(newTmpDir, rmOldTempDir = FALSE)
}
\arguments{
\item{newTmpDir}{character, indicating the path to a new temporary directory.
If the directory does not exists, it will be created.}

\item{rmOldTempDir}{logical, indicating whether to delete the currently set temp directory.}
}
\value{
character, indicating the path of the new session temporary directory.
}
\description{
This can be useful under a few circumstances:
\enumerate{
\item when performing large file operations using \code{tempdir()} or \code{tempfile()},
which may fill up the system's temporary storage (e.g., \verb{/tmp}), which could hang the system;
\item immediately after forking the R process, to ensure multiple processes do not conflict.
}
}
\note{
Use at the start of a new R session (or immediately after forking) to switch
the temporary directory location used for the session (or forked process).
}
\examples{
tempdir() ## current temp dir
myScratchDir <- file.path(dirname(tempdir()), "scratch") ## new temp dir to use
setTmpDir(myScratchDir, rmOldTempDir = TRUE)
tempdir() ## current tempdir has been updated
unlink(myScratchDir, recursive = TRUE)
}
