% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tmpdir-package.R
\docType{package}
\name{tmpdir-package}
\alias{tmpdir}
\alias{tmpdir-package}
\title{Set the location of the R session's temporary directory}
\description{
This can be useful under a few circumstances:
\enumerate{
\item when performing large file operations using \code{tempdir()} or \code{tempfile()},
which may fill up the system's temporary storage (e.g., \verb{/tmp}), which could hang the system;
\item immediately after forking the R process, to ensure multiple processes do not conflict.
}
}
\note{
Use at the start of a new R session (or immediately after forking) to switch
the temporary directory location used for the session (or forked process).
}
\author{
\strong{Maintainer}: Alex M Chubaty \email{achubaty@for-cast.ca} (\href{https://orcid.org/0000-0001-7146-8135}{ORCID})

}
